@@ -3,62 +3,37 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "prepare": "husky install",
-    "dev": "vue-cli-service serve",
-    "build": "vue-cli-service build",
+    "dev": "vite",
+    "build": "vue-tsc --noEmit && vite build",
+    "preview": "vite preview",
     "lint": "eslint . --ext .vue,.js,.jsx,.cjs,.mjs,.ts,.tsx,.cts,.mts --fix --ignore-path .gitignore",
-    "format": "prettier --write \"**/*.{js,vue,json,md}\"",
-    "pre-commit": "lint-staged"
-  },
-  "lint-staged": {
-    "*.{js,jsx,vue}": [
-      "eslint --fix",
-      "prettier --write"
-    ],
-    "*.{json,md}": [
-      "prettier --write"
-    ]
+    "format": "prettier --write \"**/*.{js,vue,json,md,ts,tsx}\""
   },
   "dependencies": {
+    "@element-plus/icons-vue": "^2.1.0",
     "ace-builds": "^1.12.3",
-    "core-js": "^3.8.3",
     "echarts": "^5.4.1",
-    "element-ui": "^2.14.1",
+    "element-plus": "^2.3.6",
     "html-to-image": "^1.9.0",
     "mathjs": "^10.5.3",
+    "mitt": "^3.0.0",
     "nanoid": "^4.0.0",
-    "vue": "^2.7.14",
-    "vue-echarts": "^6.5.4",
-    "vue-router": "^3.5.1",
-    "vuex": "^3.6.2"
+    "pinia": "^2.1.3",
+    "vue": "^3.3.4",
+    "vue-echarts": "^6.6.0",
+    "vue-router": "^4.2.2"
   },
   "devDependencies": {
-    "@babel/core": "^7.12.16",
-    "@babel/eslint-parser": "^7.12.16",
-    "@vue/cli-plugin-babel": "~5.0.0",
-    "@vue/cli-plugin-eslint": "~5.0.0",
-    "@vue/cli-plugin-router": "~5.0.0",
-    "@vue/cli-plugin-vuex": "~5.0.0",
-    "@vue/cli-service": "~5.0.0",
-    "@vue/eslint-config-airbnb": "^6.0.0",
-    "chalk": "^4.0.0",
-    "compression-webpack-plugin": "^6.1.1",
-    "eslint": "^8.0.0",
-    "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-import": "^2.25.3",
-    "eslint-plugin-prettier": "^5.2.1",
-    "eslint-plugin-vue": "^8.0.3",
-    "eslint-plugin-vuejs-accessibility": "^1.1.0",
-    "husky": "^8.0.1",
-    "lint-staged": "^11.0.0",
-    "postcss-html": "^1.5.0",
-    "postcss-scss": "^4.0.4",
-    "prettier": "npm:@btmills/prettier@^3.3.3",
-    "sass": "^1.32.7",
-    "sass-loader": "^12.0.0",
-    "stylelint": "^14.9.1",
-    "stylelint-config-recommended-vue": "^1.4.0",
-    "stylelint-config-standard-scss": "^5.0.0",
-    "vue-template-compiler": "^2.6.14"
+    "@types/node": "^20.3.1",
+    "@vitejs/plugin-vue": "^4.2.3",
+    "sass": "^1.63.6",
+    "typescript": "^5.0.2",
+    "vite": "^4.3.9",
+    "vue-tsc": "^1.8.0",
+    "eslint": "^8.42.0",
+    "eslint-config-prettier": "^8.8.0",
+    "eslint-plugin-prettier": "^4.2.1",
+    "eslint-plugin-vue": "^9.14.1",
+    "prettier": "^2.8.8"
   }
 }
