@@ -3,33 +3,33 @@
     <p class="title">画布属性</p>
     <el-form style="padding: 20px">
       <el-form-item v-for="(key, index) in Object.keys(options)" :key="index" :label="options[key]">
-        <el-color-picker v-if="isIncludesColor(key)" v-model="canvasStyleData[key]" show-alpha></el-color-picker>
+        <el-color-picker
+          v-if="isIncludesColor(key)"
+          v-model="canvasStyleData[key]"
+          show-alpha
+        ></el-color-picker>
         <el-input v-else v-model.number="canvasStyleData[key]" type="number" />
       </el-form-item>
     </el-form>
   </div>
 </template>
 
-<script>
-import { mapState } from 'vuex'
+<script setup lang="ts">
+import { useMainStore } from '@/store'
+import { storeToRefs } from 'pinia'
 
-export default {
-  data() {
-    return {
-      options: {
-        color: '颜色',
-        opacity: '不透明度',
-        backgroundColor: '背景色',
-        fontSize: '字体大小',
-      },
-    }
-  },
-  computed: mapState(['canvasStyleData']),
-  methods: {
-    isIncludesColor(str) {
-      return str.toLowerCase().includes('color')
-    },
-  },
+const store = useMainStore()
+const { canvasStyleData } = storeToRefs(store)
+
+const options: any = {
+  color: '颜色',
+  opacity: '不透明度',
+  backgroundColor: '背景色',
+  fontSize: '字体大小',
+}
+
+const isIncludesColor = (str: string) => {
+  return str.toLowerCase().includes('color')
 }
 </script>
 
