@@ -20,68 +20,52 @@
   </div>
 </template>
 
-<script>
-import { mapState } from 'vuex'
+<script setup lang="ts">
+import { useMainStore } from '@/store'
+import { storeToRefs } from 'pinia'
 
-export default {
-  data() {
-    return {
-      copyData: null,
-    }
-  },
-  computed: mapState(['menuTop', 'menuLeft', 'menuShow', 'curComponent']),
-  methods: {
-    lock() {
-      this.$store.commit('lock')
-    },
+const store = useMainStore()
+const { menuTop, menuLeft, menuShow, curComponent } = storeToRefs(store)
 
-    unlock() {
-      this.$store.commit('unlock')
-    },
+const lock = () => store.lock()
 
-    // 点击菜单时不取消当前组件的选中状态
-    handleMouseUp() {
-      this.$store.commit('setClickComponentStatus', true)
-    },
+const unlock = () => store.unlock()
 
-    cut() {
-      this.$store.commit('cut')
-    },
+// 点击菜单时不取消当前组件的选中状态
+const handleMouseUp = () => store.setClickComponentStatus(true)
 
-    copy() {
-      this.$store.commit('copy')
-    },
+const cut = () => store.cut()
 
-    paste() {
-      this.$store.commit('paste', true)
-      this.$store.commit('recordSnapshot')
-    },
+const copy = () => store.copy()
 
-    deleteComponent() {
-      this.$store.commit('deleteComponent')
-      this.$store.commit('recordSnapshot')
-    },
+const paste = () => {
+  store.paste(true)
+  store.recordSnapshot()
+}
 
-    upComponent() {
-      this.$store.commit('upComponent')
-      this.$store.commit('recordSnapshot')
-    },
+const deleteComponent = () => {
+  store.deleteComponent()
+  store.recordSnapshot()
+}
 
-    downComponent() {
-      this.$store.commit('downComponent')
-      this.$store.commit('recordSnapshot')
-    },
+const upComponent = () => {
+  store.upComponent()
+  store.recordSnapshot()
+}
 
-    topComponent() {
-      this.$store.commit('topComponent')
-      this.$store.commit('recordSnapshot')
-    },
+const downComponent = () => {
+  store.downComponent()
+  store.recordSnapshot()
+}
 
-    bottomComponent() {
-      this.$store.commit('bottomComponent')
-      this.$store.commit('recordSnapshot')
-    },
-  },
+const topComponent = () => {
+  store.topComponent()
+  store.recordSnapshot()
+}
+
+const bottomComponent = () => {
+  store.bottomComponent()
+  store.recordSnapshot()
 }
 </script>
 
