@@ -18,52 +18,51 @@
   </div>
 </template>
 
-<script>
-import { mapState } from 'vuex'
+<script setup lang="ts">
+import { useMainStore } from '@/store'
+import { storeToRefs } from 'pinia'
 
-export default {
-  computed: mapState(['componentData', 'curComponent', 'curComponentIndex']),
-  methods: {
-    getComponent(index) {
-      return this.componentData[this.componentData.length - 1 - index]
-    },
+const store = useMainStore()
+const { componentData, curComponentIndex } = storeToRefs(store)
 
-    transformIndex(index) {
-      return this.componentData.length - 1 - index
-    },
+const getComponent = (index: number) => {
+  return componentData.value[componentData.value.length - 1 - index]
+}
 
-    onClick(index) {
-      if (!this.$store.state.rightList) {
-        this.$store.commit('isShowRightList')
-      }
-      this.setCurComponent(index)
-    },
+const transformIndex = (index: number) => {
+  return componentData.value.length - 1 - index
+}
 
-    deleteComponent() {
-      setTimeout(() => {
-        this.$store.commit('deleteComponent')
-        this.$store.commit('recordSnapshot')
-      })
-    },
+const onClick = (index: number) => {
+  if (!store.rightList) {
+    store.isShowRightList()
+  }
+  setCurComponent(index)
+}
 
-    upComponent() {
-      setTimeout(() => {
-        this.$store.commit('upComponent')
-        this.$store.commit('recordSnapshot')
-      })
-    },
+const deleteComponent = (index: number) => {
+  setTimeout(() => {
+    store.deleteComponent(index)
+    store.recordSnapshot()
+  })
+}
 
-    downComponent() {
-      setTimeout(() => {
-        this.$store.commit('downComponent')
-        this.$store.commit('recordSnapshot')
-      })
-    },
+const upComponent = (index: number) => {
+  setTimeout(() => {
+    store.upComponent()
+    store.recordSnapshot()
+  })
+}
 
-    setCurComponent(index) {
-      this.$store.commit('setCurComponent', { component: this.componentData[index], index })
-    },
-  },
+const downComponent = (index: number) => {
+  setTimeout(() => {
+    store.downComponent()
+    store.recordSnapshot()
+  })
+}
+
+const setCurComponent = (index: number) => {
+  store.setCurComponent({ component: componentData.value[index], index })
 }
 </script>
 
