@@ -3,12 +3,7 @@
     <table @dblclick="onDblclick">
       <tbody>
         <tr v-for="(item, row) in tableData" :key="row">
-          <td
-            v-for="(e, col) in item"
-            :key="col"
-            :class="{ selected: curTd === row + ',' + col }"
-            @click="onClick(row, col)"
-          >
+          <td v-for="(e, col) in item" :key="col" :class="{ selected: curTd === row + ',' + col }" @click="onClick(row, col)">
             <el-input
               v-if="canEdit && curTd === row + ',' + col"
               v-model="tableData[row][col]"
@@ -20,88 +15,81 @@
         </tr>
       </tbody>
     </table>
-    <div>
-      <button @click="addRow">添加新行</button>
-      <button @click="addCol">添加新列</button>
-      <button @click="deleteRow">删除行</button>
-      <button @click="deleteCol">删除列</button>
+    <div style="margin-top: 10px">
+      <el-button @click="addRow">添加行</el-button>
+      <el-button @click="addCol">添加列</el-button>
+      <el-button @click="deleteRow">删除行</el-button>
+      <el-button @click="deleteCol">删除列</el-button>
     </div>
   </div>
 </template>
 
-<script>
-export default {
-  directives: {
-    focus: {
-      // 指令的定义
-      inserted(el) {
-        // 聚焦元素
-        el.querySelector('input').focus()
-      },
-    },
-  },
-  data() {
-    return {
-      curProperty: '',
-      curTd: '',
-      canEdit: false,
-      preCurTd: '', // 失焦时 curTd 值为空，这时删除会读不到值，因此用这个变量来代替，用于删除行列
-    }
-  },
-  computed: {
-    tableData() {
-      return this.$store.state.curComponent.propValue.data
-    },
-  },
-  methods: {
-    onDblclick() {
-      this.canEdit = true
-    },
+<script setup lang="ts">
+import { ref, computed } from 'vue'
+import { useMainStore } from '@/store'
+import { storeToRefs } from 'pinia'
+import toast from '@/utils/toast'
 
-    onClick(row, col) {
-      this.curTd = `${row},${col}`
-      this.preCurTd = this.curTd
-    },
+const store = useMainStore()
+const { curComponent } = storeToRefs(store)
 
-    onBlur() {
-      this.canEdit = false
-      this.curTd = ''
-    },
+const curTd = ref('')
+const canEdit = ref(false)
 
-    deleteRow() {
-      if (!this.preCurTd) {
-        this.$message.error('请先选择要删除的行')
-        return
-      }
+const tableData = computed(() => {
+  return curComponent.value.propValue.data
+})
 
-      const row = this.preCurTd.split(',')[0]
-      this.tableData.splice(row, 1)
-    },
+const onDblclick = () => {
+  canEdit.value = true
+}
 
-    addRow() {
-      this.tableData.push(new Array(this.tableData[0]?.length || 1).fill(' '))
-    },
+const onClick = (row: number, col: number) => {
+  curTd.value = `${row},${col}`
+}
 
-    addCol() {
-      if (this.tableData.length) {
-        this.tableData.forEach((item) => item.push(' '))
-      } else {
-        this.tableData.push([' '])
-      }
-    },
+const onBlur = () => {
+  canEdit.value = false
+  curTd.value = ''
+}
 
-    deleteCol() {
-      if (!this.preCurTd) {
-        this.$message.error('请先选择要删除的列')
-        return
-      }
+const deleteRow = () => {
+  if (!curTd.value) {
+    toast('请先选择单元格')
+    return
+  }
 
-      const col = this.preCurTd.split(',')[1]
-      this.tableData.forEach((item) => {
-        item.splice(col, 1)
-      })
-    },
-  },
+  const row = Number(curTd.value.split(',')[0])
+  tableData.value.splice(row, 1)
+}
+
+const addRow = () => {
+  const cols = tableData.value[0].length
+  const newRow = new Array(cols).fill(' ')
+  tableData.value.push(newRow)
+}
+
+const addCol = () => {
+  tableData.value.forEach((item: any) => {
+    item.push(' ')
+  })
+}
+
+const deleteCol = () => {
+  if (!curTd.value) {
+    toast('请先选择单元格')
+    return
+  }
+
+  const col = Number(curTd.value.split(',')[1])
+  tableData.value.forEach((item: any) => {
+    item.splice(col, 1)
+  })
+}
+
+// v-focus directive
+const vFocus = {
+  mounted: (el: HTMLElement) => el.querySelector('input')?.focus(),
 }
 </script>
 
@@ -110,53 +98,34 @@ export default {
   overflow: auto;
   margin-bottom: 8px;
 
-  & > div {
-    margin-top: 18px;
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: space-between;
-
-    button {
-      cursor: pointer;
-      background: #fff;
-      border: 1px solid #dcdfe6;
-      color: #606266;
-      text-align: center;
-      box-sizing: border-box;
-      outline: 0;
-      margin: 0;
-      font-weight: 500;
-      padding: 4px 10px;
-      font-size: 14px;
-      border-radius: 4px;
-      margin-bottom: 10px;
-
-      &:hover {
-        background: #ecf5ff;
-        color: #409eff;
-      }
-    }
-  }
-
-  table {
+  & > table {
     border-collapse: collapse;
+    width: 100%;
+    margin-bottom: 10px;
     word-break: break-all;
     word-wrap: break-word;
-    text-align: center;
-    font-size: 12px;
 
     td {
       border: 1px solid #ebeef5;
       height: 40px;
       min-width: 60px;
-      max-width: 80px;
-      padding: 10px;
+      text-align: center;
+      cursor: pointer;
+      box-sizing: border-box;
+      position: relative;
+
+      &:hover {
+        background-color: #f5f7fa;
+      }
+    }
+
+    .selected {
+      border: 2px solid #409eff;
     }
   }
 
-  .selected {
-    background: #ecf5ff;
-    color: #409eff;
+  .el-button {
+    margin-bottom: 10px;
   }
 }
 </style>
