@@ -6,7 +6,7 @@
       <div>
         <el-tag v-for="(tag, index) in curComponent.animations" :key="index" closable @close="removeAnimation(index)">
           {{ tag.label }}
-          <i class="cursor el-icon-setting" @click="handleAnimationSetting(index)"></i>
+          <el-icon class="cursor" @click="handleAnimationSetting(index)"><Setting /></el-icon>
         </el-tag>
       </div>
     </div>
@@ -18,10 +18,10 @@
           <el-scrollbar class="animate-container">
             <div
               v-for="animate in item.children"
-              :ref="animate.value"
+              :ref="(el) => setAnimateRef(el, animate.value)"
               :key="animate.value"
               class="animate"
-              @mouseenter="runAnimation(animate)"
+              @mouseenter="runAnimationFn(animate)"
               @click="addAnimation(animate)"
             >
               <div>
@@ -41,64 +41,64 @@
   </div>
 </template>
 
-<script>
-import Modal from '@/components/Modal'
+<script setup lang="ts">
+import { ref } from 'vue'
+import { useMainStore } from '@/store'
+import { storeToRefs } from 'pinia'
+import Modal from '@/components/Modal.vue'
 import eventBus from '@/utils/eventBus'
 import animationClassData from '@/utils/animationClassData'
-import { mapState } from 'vuex'
 import runAnimation from '@/utils/runAnimation'
 import AnimationSettingModal from './AnimationSettingModal.vue'
+import { Setting } from '@element-plus/icons-vue'
 
-export default {
-  components: { Modal, AnimationSettingModal },
-  data() {
-    return {
-      isShowAnimation: false,
-      hoverPreviewAnimate: '',
-      animationActiveName: '进入',
-      animationClassData,
-      showAnimatePanel: false,
-      timer: null,
-      isShowAnimationSetting: false,
-      curIndex: 0,
-    }
-  },
-  computed: mapState(['curComponent']),
-  methods: {
-    addAnimation(animate) {
-      this.$store.commit('addAnimation', animate)
-      this.isShowAnimation = false
-    },
+const store = useMainStore()
+const { curComponent } = storeToRefs(store)
 
-    previewAnimate() {
-      eventBus.$emit('runAnimation')
-    },
+const isShowAnimation = ref(false)
+const animationActiveName = ref('进入')
+const isShowAnimationSetting = ref(false)
+const curIndex = ref(0)
+const animateRefs = ref<any>({})
 
-    removeAnimation(index) {
-      this.$store.commit('removeAnimation', index)
-      if (!this.curComponent.animations.length) {
-        // 清空动画数据，停止运动
-        eventBus.$emit('stopAnimation')
-      }
-    },
+const setAnimateRef = (el: any, value: string) => {
+  if (el) {
+    animateRefs.value[value] = el
+  }
+}
 
-    handleAnimationSetting(index) {
-      this.isShowAnimationSetting = true
-      this.curIndex = index
-    },
+const addAnimation = (animate: any) => {
+  store.addAnimation(animate)
+  isShowAnimation.value = false
+}
 
-    async runAnimation(animate) {
-      if (animate.pending) return
+const previewAnimate = () => {
+  eventBus.emit('runAnimation')
+}
 
-      animate.pending = true
-      await runAnimation(this.$refs[animate.value][0], [animate])
+const removeAnimation = (index: number) => {
+  store.removeAnimation(index)
+  if (!curComponent.value.animations.length) {
+    // 清空动画数据，停止运动
+    eventBus.emit('stopAnimation')
+  }
+}
 
-      // 防止无限触发同一元素的动画
-      setTimeout(() => {
-        animate.pending = false
-      }, 100)
-    },
-  },
+const handleAnimationSetting = (index: number) => {
+  isShowAnimationSetting.value = true
+  curIndex.value = index
+}
+
+const runAnimationFn = async (animate: any) => {
+  if (animate.pending) return
+
+  animate.pending = true
+  await runAnimation(animateRefs.value[animate.value], [animate])
+
+  // 防止无限触发同一元素的动画
+  setTimeout(() => {
+    animate.pending = false
+  }, 100)
 }
 </script>
 
@@ -106,6 +106,7 @@ export default {
 .animation-list {
   .cursor {
     cursor: pointer;
+    margin-left: 5px;
   }
 
   .div-animation {
@@ -116,7 +117,9 @@ export default {
     }
 
     .el-tag {
-      display: block;
+      display: flex;
+      align-items: center;
+      justify-content: center;
       width: 50%;
       margin: auto;
       margin-bottom: 10px;
