@@ -1,11 +1,11 @@
 <template>
   <div class="ace">
     <div class="header">
-      <el-button class="btn" icon="el-icon-search" @click="openSearchBox"> 查找 </el-button>
-      <el-button class="btn" icon="el-icon-close" @click="closeEditor"> 关闭 </el-button>
+      <el-button class="btn" :icon="Search" @click="openSearchBox"> 查找 </el-button>
+      <el-button class="btn" :icon="Close" @click="closeEditor"> 关闭 </el-button>
     </div>
     <div class="ace-editor">
-      <div ref="ace" class="editor" />
+      <div ref="editorRef" class="editor" />
     </div>
     <div class="footer">
       <el-button type="primary" @click="setCode"> 重置代码 </el-button>
@@ -14,34 +14,39 @@
   </div>
 </template>
 
-<script>
+<script setup lang="ts">
+import { ref, onMounted, watch } from 'vue'
 import ace from 'ace-builds'
 import 'ace-builds/src-min-noconflict/theme-one_dark'
 import 'ace-builds/src-min-noconflict/ext-searchbox'
 import 'ace-builds/src-min-noconflict/mode-json5'
 import 'ace-builds/src-min-noconflict/ext-language_tools'
-import { mapState } from 'vuex'
+import { useMainStore } from '@/store'
+import { storeToRefs } from 'pinia'
+import { Search, Close } from '@element-plus/icons-vue'
 
-export default {
-  name: 'AceEditor',
-  data() {
-    return {
-      editor: null,
-      obj: null,
-    }
-  },
-  computed: mapState(['canvasStyleData', 'curComponent']),
-  watch: {
-    curComponent() {
-      this.setCode()
-    },
-    canvasStyleData() {
-      this.setCode()
-    },
-  },
-  mounted() {
-    ace.config.set('basePath', 'https://cdnjs.cloudflare.com/ajax/libs/ace/1.4.14/')
-    this.editor = ace.edit(this.$refs.ace, {
+const emit = defineEmits(['closeEditor'])
+
+const store = useMainStore()
+const { canvasStyleData, curComponent } = storeToRefs(store)
+
+const editorRef = ref<HTMLElement | null>(null)
+let editor: any = null
+const obj = ref<any>(null)
+
+const setCode = () => {
+  obj.value = curComponent.value || canvasStyleData.value
+  editor.setValue(JSON.stringify(obj.value, null, 4))
+}
+
+watch([curComponent, canvasStyleData], () => {
+  setCode()
+})
+
+onMounted(() => {
+  ace.config.set('basePath', 'https://cdnjs.cloudflare.com/ajax/libs/ace/1.4.14/')
+  if (editorRef.value) {
+    editor = ace.edit(editorRef.value, {
       maxLines: 40,
       minLines: 40,
       fontSize: 14,
@@ -54,36 +59,25 @@ export default {
       enableSnippets: true,
     })
 
-    this.obj = this.curComponent || this.canvasStyleData
-    this.editor.setValue(JSON.stringify(this.obj, null, 4))
-  },
-  methods: {
-    setCode() {
-      this.obj = this.curComponent || this.canvasStyleData
-      this.editor.setValue(JSON.stringify(this.obj, null, 4))
-    },
+    setCode()
+  }
+})
 
-    getCode() {
-      let str = this.editor.getValue()
-      if (!this.curComponent) {
-        this.$store.commit('aceSetCanvasData', JSON.parse(str))
-      } else {
-        this.$store.commit('aceSetcurComponent', JSON.parse(str))
-      }
-    },
+const getCode = () => {
+  let str = editor.getValue()
+  if (!curComponent.value) {
+    store.aceSetCanvasData(JSON.parse(str))
+  } else {
+    store.aceSetcurComponent(JSON.parse(str))
+  }
+}
 
-    updateEditorTheme(theme) {
-      this.editor.setTheme(theme)
-    },
+const openSearchBox = () => {
+  editor.execCommand('find')
+}
 
-    openSearchBox() {
-      this.editor.execCommand('find')
-    },
-
-    closeEditor() {
-      this.$emit('closeEditor')
-    },
-  },
+const closeEditor = () => {
+  emit('closeEditor')
 }
 </script>
 
