@@ -4,23 +4,19 @@
     <el-form>
       <el-form-item label="镜像翻转">
         <div style="clear: both">
-          <el-checkbox v-model="curComponent.propValue.flip.horizontal" label="horizontal">水平翻转</el-checkbox>
-          <el-checkbox v-model="curComponent.propValue.flip.vertical" label="vertical">垂直翻转</el-checkbox>
+          <el-checkbox v-model="curComponent.propValue.flip.horizontal" label="水平翻转"></el-checkbox>
+          <el-checkbox v-model="curComponent.propValue.flip.vertical" label="垂直翻转"></el-checkbox>
         </div>
       </el-form-item>
     </el-form>
   </div>
 </template>
 
-<script>
+<script setup lang="ts">
 import CommonAttr from '@/custom-component/common/CommonAttr.vue'
+import { useMainStore } from '@/store'
+import { storeToRefs } from 'pinia'
 
-export default {
-  components: { CommonAttr },
-  computed: {
-    curComponent() {
-      return this.$store.state.curComponent
-    },
-  },
-}
+const store = useMainStore()
+const { curComponent } = storeToRefs(store)
 </script>
