@@ -3,21 +3,17 @@
     <CommonAttr></CommonAttr>
     <el-form>
       <el-form-item label="内容">
-        <el-input v-model="curComponent.propValue" type="textarea" :rows="3" />
+        <el-input v-model="curComponent.propValue" />
       </el-form-item>
     </el-form>
   </div>
 </template>
 
-<script>
+<script setup lang="ts">
 import CommonAttr from '@/custom-component/common/CommonAttr.vue'
+import { useMainStore } from '@/store'
+import { storeToRefs } from 'pinia'
 
-export default {
-  components: { CommonAttr },
-  computed: {
-    curComponent() {
-      return this.$store.state.curComponent
-    },
-  },
-}
+const store = useMainStore()
+const { curComponent } = storeToRefs(store)
 </script>
