@@ -1,28 +1,26 @@
 <template>
-  <div class="attr-list v-talbe-attr">
+  <div class="attr-list">
     <CommonAttr></CommonAttr>
-    <EditTable></EditTable>
     <el-form>
       <el-form-item label="斑马纹">
-        <el-switch v-model="propValue.stripe"></el-switch>
+        <el-switch v-model="curComponent.propValue.stripe"></el-switch>
       </el-form-item>
       <el-form-item label="表头加粗">
-        <el-switch v-model="propValue.thBold"></el-switch>
+        <el-switch v-model="curComponent.propValue.thBold"></el-switch>
+      </el-form-item>
+      <el-form-item label="表格数据">
+        <EditTable></EditTable>
       </el-form-item>
     </el-form>
   </div>
 </template>
 
-<script>
-import CommonAttr from '@/custom-component/common/CommonAttr'
-import EditTable from './EditTable'
+<script setup lang="ts">
+import CommonAttr from '@/custom-component/common/CommonAttr.vue'
+import EditTable from './EditTable.vue'
+import { useMainStore } from '@/store'
+import { storeToRefs } from 'pinia'
 
-export default {
-  components: { EditTable, CommonAttr },
-  computed: {
-    propValue() {
-      return this.$store.state.curComponent.propValue
-    },
-  },
-}
+const store = useMainStore()
+const { curComponent } = storeToRefs(store)
 </script>
