@@ -1,21 +1,22 @@
 import { deepCopy } from '@/utils/utils'
-import store from '@/store'
+import { useMainStore } from '@/store'
 import { divide, multiply } from 'mathjs'
 
 const needToChangeAttrs = ['top', 'left', 'width', 'height', 'fontSize', 'padding']
 // 根据比例缩放组件尺寸
-export default function changeComponentsSizeWithScale(scale, snapshotData = null) {
-  const componentData = snapshotData || deepCopy(store.state.componentData)
-  componentData.forEach((component) => {
+export default function changeComponentsSizeWithScale(scale: number, snapshotData: any = null) {
+  const store = useMainStore()
+  const componentData = snapshotData || deepCopy(store.componentData)
+  componentData.forEach((component: any) => {
     Object.keys(component.style).forEach((key) => {
       if (needToChangeAttrs.includes(key)) {
         let newKey
         if (snapshotData) {
           // 根据比例计算新的属性值
-          newKey = ((component.style[key] / snapshotData[0].lastScale) * scale).toFixed(4) - 0
+          newKey = Number(((component.style[key] / snapshotData[0].lastScale) * scale).toFixed(4))
         } else {
           // 否则根据当前画布的比例计算新的属性值
-          newKey = ((component.style[key] / store.state.canvasStyleData.scale) * scale).toFixed(4) - 0
+          newKey = Number(((component.style[key] / store.canvasStyleData.scale) * scale).toFixed(4))
         }
 
         if (key == 'top' || key == 'left') {
@@ -31,31 +32,32 @@ export default function changeComponentsSizeWithScale(scale, snapshotData = null
     return componentData
   }
 
-  store.commit('setComponentData', componentData)
+  store.setComponentData(componentData)
   // 更新后的组件数据
-  store.commit('setCurComponent', {
-    component: componentData[store.state.curComponentIndex],
-    index: store.state.curComponentIndex,
+  store.setCurComponent({
+    component: componentData[store.curComponentIndex as number],
+    index: store.curComponentIndex,
   })
 
   // 更新画布的比例
-  store.commit('setCanvasStyle', {
-    ...store.state.canvasStyleData,
+  store.setCanvasStyle({
+    ...store.canvasStyleData,
     scale,
   })
 }
 
 const needToChangeAttrs2 = ['width', 'height', 'fontSize', 'padding']
-export function changeComponentSizeWithScale(component) {
+export function changeComponentSizeWithScale(component: any) {
+  const store = useMainStore()
   Object.keys(component.style).forEach((key) => {
     if (needToChangeAttrs2.includes(key)) {
       if (key === 'fontSize' && component.style[key] === '') return
 
-      component.style[key] = format(component.style[key], store.state.canvasStyleData.scale)
+      component.style[key] = format(component.style[key], store.canvasStyleData.scale)
     }
   })
 }
 
-function format(value, scale) {
-  return multiply(value, divide(parseFloat(scale), 100))
+function format(value: any, scale: number) {
+  return multiply(value, divide(scale, 100))
 }
